# Precise Rib Parsing & Errors