--- a/golem-rib/src/parser/flag.rs
+++ b/golem-rib/src/parser/flag.rs
@@ -21,21 +21,24 @@ use combine::{
 
 use crate::expr::Expr;
 use combine::sep_by;
-use combine::stream::easy;
 
-pub fn flag<'t>() -> impl Parser<easy::Stream<&'t str>, Output = Expr> {
+pub fn flag<Input>() -> impl Parser<Input, Output = Expr>
+where
+    Input: combine::Stream<Token = char>,
+{
     let flag_name = many1(letter().or(char_('_')).or(digit()).or(char_('-')))
         .map(|s: Vec<char>| s.into_iter().collect());
 
-    spaces().with(
-        between(
-            char_('{').skip(spaces()),
-            char_('}').skip(spaces()),
-            sep_by(flag_name.skip(spaces()), char_(',').skip(spaces())),
+    spaces()
+        .with(
+            between(
+                char_('{').skip(spaces()),
+                char_('}').skip(spaces()),
+                sep_by(flag_name.skip(spaces()), char_(',').skip(spaces())),
+            )
+            .map(Expr::flags),
         )
-        .map(Expr::flags)
-        .message("Unable to parse flag"),
-    )
+        .message("Invalid syntax for flag type")
 }
 
 #[cfg(test)]